Add _rosenbrock parent class
